requirements.txt
mysql-connector-python==8.4.0
matplotlib==3.8.4
tkcalendar==1.6.1
Pillow==10.3.0
